+++ docker-compose.yml
@@ -7,7 +7,7 @@ services:
       dockerfile: jarvis.Dockerfile
     command: jarvis
     ports:
-      - 3030:3030
+      - 3000:3000
     environment:
       - WEB_INTERFACE=0.0.0.0
       - WEB_PORT=3000
