+++ infrastructure/controllers/app_controller_test.go
@@ -14,17 +14,6 @@ type AppControllerSuite struct {
 }
 
 func (suite *AppControllerSuite) SetupTest() {
-	suite.AppController = NewAppController()
-	suite.AppController.Setup()
-
-	suite.WebSuite.Server = suite.AppController.Engine.Server
-	suite.Suite.SetupTest()
-}
-
-func (suite *AppControllerSuite) TestNewAppController() {
-	suite.NotEmpty(suite.AppController)
-	suite.NotEmpty(suite.AppController.RootController)
-	suite.NotEmpty(suite.AppController.RootController.AppController)
 }
 
 func TestAppControllerSuite(t *testing.T) {
