+++ infrastructure/controllers/root_controller.go
@@ -4,27 +4,26 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	"github.com/magrathealabs/jarvis/infrastructure/routes"
-	"github.com/magrathealabs/jarvis/infrastructure/web"
 )
 
 // RootController server get at RootPath
 type RootController struct {
-	Engine        *web.Engine
 	AppController *AppController
 }
 
 // NewRootController costructor
 func NewRootController(other *AppController) *RootController {
-	return &RootController{
+	controller := &RootController{
 		AppController: other,
-		Engine:        other.Engine,
 	}
+
+	controller.Routes()
+
+	return controller
 }
 
-// Setup this controller
-func (controller *RootController) Setup() {
-	controller.Engine.Server.GET(routes.Root, controller.Index)
+// Routes setup all handlers
+func (controller *RootController) Routes() {
 }
 
 // Index GET routes.Root
