+++ infrastructure/controllers/root_controller_test.go
@@ -1,10 +1,7 @@
 package controllers
 
 import (
-	"net/http"
 	"testing"
-
-	"github.com/magrathealabs/jarvis/infrastructure/routes"
 
 	"github.com/krakenlab/gspec"
 	"github.com/krakenlab/gspec/websuite"
@@ -18,17 +15,6 @@ type RootControllerSuite struct {
 }
 
 func (suite *RootControllerSuite) SetupTest() {
-	suite.AppController = NewAppController()
-	suite.AppController.Setup()
-
-	suite.RootController = suite.AppController.RootController
-	suite.WebSuite.Server = suite.AppController.Engine.Server
-	suite.Suite.SetupTest()
-}
-
-func (suite *RootControllerSuite) TestIndex() {
-	response := suite.GET(routes.Root, "")
-	suite.Equal(http.StatusOK, response.Code)
 }
 
 func TestRootControllerSuite(t *testing.T) {
